chore: rename networkRequest -> NetworkRequest
