+++ src/bidiMapper/domains/network/NetworkRequest.ts
